feedback: move to util
