+++ packages/sdk-browser/src/sdk-client.ts
@@ -1,14 +1,6 @@
 import { WebauthnStamper } from "@turnkey/webauthn-stamper";
 import { IframeStamper } from "@turnkey/iframe-stamper";
-import bs58check from "bs58check";
-import { AeadId, CipherSuite, KdfId, KemId } from "hpke-js";
-import { uint8ArrayFromHexString } from "@turnkey/encoding";
 import { getWebAuthnAttestation } from "@turnkey/http";
-import {
-  generateP256KeyPair,
-  buildAdditionalAssociatedData,
-  compressRawPublicKey,
-} from "@turnkey/crypto";
 
 import { VERSION } from "./__generated__/version";
 import WindowWrapper from "./__polyfills__/window";
@@ -25,14 +17,18 @@ import { TurnkeyRequestError } from "./__types__/base";
 import { TurnkeySDKClientBase } from "./__generated__/sdk-client-base";
 import type * as SdkApiTypes from "./__generated__/sdk_api_types";
 
-import type { User, SubOrganization, EmbeddedAPIKey } from "./models";
+import type { User, SubOrganization } from "./models";
 import {
   StorageKeys,
   getStorageValue,
   removeStorageValue,
   setStorageValue,
 } from "./storage";
-import { generateRandomBuffer, base64UrlEncode } from "./utils";
+import {
+  generateRandomBuffer,
+  base64UrlEncode,
+  createEmbeddedAPIKey,
+} from "./utils";
 
 export class TurnkeyBrowserSDK {
   config: TurnkeySDKBrowserConfig;
@@ -195,64 +191,6 @@ export class TurnkeyBrowserClient extends TurnkeySDKClientBase {
     await setStorageValue(StorageKeys.CurrentUser, currentUser);
     return readOnlySessionResult!;
   };
-
-  // createEmbeddedAPIKey creates an embedded API key encrypted to a target key (typically embedded within an iframe).
-  // This returns a bundle that can be decrypted by that target key, as well as the public key of the newly created API key.
-  createEmbeddedAPIKey = async (
-    targetPublicKey: string
-  ): Promise<EmbeddedAPIKey> => {
-    const TURNKEY_HPKE_INFO = new TextEncoder().encode("turnkey_hpke");
-
-    // 1: create new API key (to be encrypted to the targetPublicKey)
-    const p256key = generateP256KeyPair();
-
-    // 2: set up encryption
-    const suite = new CipherSuite({
-      kem: KemId.DhkemP256HkdfSha256,
-      kdf: KdfId.HkdfSha256,
-      aead: AeadId.Aes256Gcm,
-    });
-
-    // 3: import the targetPublicKey (i.e. passed in from the iframe)
-    const targetKeyBytes = uint8ArrayFromHexString(targetPublicKey);
-    const targetKey = await crypto.subtle.importKey(
-      "raw",
-      targetKeyBytes,
-      {
-        name: "ECDH",
-        namedCurve: "P-256",
-      },
-      true,
-      []
-    );
-
-    // 4: sender encrypts a message to the target key
-    const sender = await suite.createSenderContext({
-      recipientPublicKey: targetKey,
-      info: TURNKEY_HPKE_INFO,
-    });
-    const ciphertext = await sender.seal(
-      uint8ArrayFromHexString(p256key.privateKey),
-      buildAdditionalAssociatedData(new Uint8Array(sender.enc), targetKeyBytes)
-    );
-    const ciphertextUint8Array = new Uint8Array(ciphertext);
-
-    // 5: assemble bundle
-    const encappedKey = new Uint8Array(sender.enc);
-    const compressedEncappedKey = compressRawPublicKey(encappedKey);
-    const result = new Uint8Array(
-      compressedEncappedKey.length + ciphertextUint8Array.length
-    );
-    result.set(compressedEncappedKey);
-    result.set(ciphertextUint8Array, compressedEncappedKey.length);
-
-    const base58encodedBundle = bs58check.encode(result);
-
-    return {
-      authBundle: base58encodedBundle,
-      publicKey: p256key.publicKey,
-    };
-  };
 }
 
 export class TurnkeyPasskeyClient extends TurnkeyBrowserClient {
@@ -327,7 +265,7 @@ export class TurnkeyPasskeyClient extends TurnkeyBrowserClient {
     const localStorageUser = await getStorageValue(StorageKeys.CurrentUser);
     userId = userId ?? localStorageUser?.userId;
 
-    const { authBundle, publicKey } = await this.createEmbeddedAPIKey(
+    const { authBundle, publicKey } = await createEmbeddedAPIKey(
       targetEmbeddedKey
     );
